[System] Refactor check for null
- Breakdown the null class into manage able class size
